+++ frontend/src/app/tasks/components/CreateTask.tsx
@@ -9,7 +9,7 @@ import Divider from '@mui/joy/Divider'
 import { Play16Filled, RecordStop24Regular } from '@fluentui/react-icons'
 
 import { TimePicker } from '../components/TimePicker'
-import { useTaskForm } from '../hooks/useTaskForm'
+import { useTaskForm } from '../hooks/useCreateTaskForm'
 
 import { Project } from '@/domain/Project'
 import { TaskType } from '@/domain/TaskType'
